Added till 20/3/21

+++ template.cpp
@@ -7,10 +7,10 @@ using namespace std;
 #define OR ||
 #define yes cout << "YES" << endl
 #define no cout << "NO" << endl
-#define FOR(i, n) for (ll(i) = 0; (i) < (n); (i)++)
-#define FORL(i, a, n) for (ll(i) = (a); (i) <= (n); (i)++)
-#define FORR(i, a, n) for (ll(i) = (a); (i) >= (n); (i)--)
-#define FORSQ(i, a, n) for (ll(i) = (a); (i) * (i) <= (n); ++(i))
+#define FOR(i, n) for (int(i) = 0; (i) < (n); (i)++)
+#define FORL(i, a, n) for (int(i) = (a); (i) <= (n); (i)++)
+#define FORR(i, a, n) for (int(i) = (a); (i) >= (n); (i)--)
+#define FORSQ(i, a, n) for (int(i) = (a); (i) * (i) <= (n); ++(i))
 #define FOREACH(a, b) for (auto &(a) : (b))
 #define all(v) v.begin(), v.end()
 #define SORT(v) sort(ALL(v))
@@ -59,7 +59,7 @@ int main()
   cin.tie(NULL);
   cout.tie(NULL);
 
-  ll t = 1;
+  int t = 1;
   cin >> t;
   while (t--)
   {
